Updated analysis with better linear model
I have updated my analysis by cleaning the pca analysis and the linear model fit with test-train split. The does not show any correlation still, but I am thinking that a dummification of the month and day variable could change the picture a bit. Furthermore, I am thinking that the analysis may be not be well posed because fires that span multiple days are considered as different entries where they should be considered as a time series instead.